advertisingSystem/requirements.txt
amqp==5.1.1
asgiref==3.5.2
autopep8==2.0.0
billiard==3.6.4.0
boto3==1.26.3
botocore==1.29.3
celery==5.2.7
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==4.1.3
djangorestframework==3.14.0
idna==3.4
jmespath==1.0.1
kombu==5.2.4
Pillow==9.3.0
prompt-toolkit==3.0.32
pycodestyle==2.9.1
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.6
requests==2.28.1
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
urllib3==1.26.12
vine==5.0.0
wcwidth==0.2.5
